KL-163/test: move into subdirectories
