+++ src/test/java/taco/klkl/domain/category/controller/tag/TagControllerTest.java
@@ -1,4 +1,4 @@
-package taco.klkl.domain.category.controller;
+package taco.klkl.domain.category.controller.tag;
 
 import static org.hamcrest.Matchers.*;
 import static org.mockito.Mockito.*;
@@ -20,7 +20,6 @@ import org.springframework.http.MediaType;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.web.method.annotation.MethodArgumentTypeMismatchException;
 
-import taco.klkl.domain.category.controller.tag.TagController;
 import taco.klkl.domain.category.dao.SubcategoryTagRepository;
 import taco.klkl.domain.category.domain.SubcategoryTag;
 import taco.klkl.domain.category.domain.subcategory.Subcategory;
