+++ src/test/java/taco/klkl/domain/category/integration/category/CategoryIntegrationTest.java
@@ -1,4 +1,4 @@
-package taco.klkl.domain.category.integration;
+package taco.klkl.domain.category.integration.category;
 
 import static org.hamcrest.Matchers.*;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
