+++ src/test/java/taco/klkl/domain/region/integration/region/RegionIntegrationTest.java
@@ -1,4 +1,4 @@
-package taco.klkl.domain.region.integration;
+package taco.klkl.domain.region.integration.region;
 
 import static org.hamcrest.Matchers.*;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
